Ajout du système de connexion + Connexion BDD

+++ database/migrations/2021_02_01_165039_partners.php
@@ -16,7 +16,7 @@ class Partners extends Migration
         Schema::create('partners', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->bigIncrements('volunteers_max_score');
+            $table->integer('volunteers_max_score');
             $table->string('siret')->nullable();
             $table->string('naf')->nullable();
             $table->string('phone');
